fix(core): unable to use the same stack name for two stacks in different environments
Since we used the stack name as the template file name, if users wanted to use the
same stack name in two environments, the emitted templates overwrote each other.

This fix changes the template file name to use the `uniqueId` of the stack construct
instead of the stack name.

Fixes #4412